+++ hybrid_core/hybrid_core_darwin/lib/src/ffi.dart
@@ -2,7 +2,7 @@ import 'dart:ffi';
 
 import 'ffi.g.dart';
 
-const foundationPath =
+const _foundationPath =
     '/System/Library/Frameworks/Foundation.framework/Foundation';
-final foundationDyLib = DynamicLibrary.open(foundationPath);
-final foundationLib = HybridCoreDarwinLibrary(foundationDyLib);
+final _foundationLib = DynamicLibrary.open(_foundationPath);
+final foundationLib = HybridCoreDarwinLibrary(_foundationLib);