+++ hybrid_core/hybrid_core_linux/lib/src/core.dart
@@ -1,0 +1,9 @@
+import 'package:hybrid_core_platform_interface/hybrid_core_platform_interface.dart';
+
+import 'os.dart';
+
+abstract class HybridCoreLinuxPlugin {
+  static void registerWith() {
+    OSImpl.instance = LinuxImpl();
+  }
+}