+++ hybrid_core/hybrid_core_windows/lib/src/ffi.dart
@@ -4,8 +4,8 @@ import 'dart:ffi';
 
 import 'ffi.g.dart';
 
-final kernel32DyLib = DynamicLibrary.open('Kernel32.dll');
-final kernel32Lib = HybridCoreWindowsLibrary(kernel32DyLib);
+final _kernel32Lib = DynamicLibrary.open('Kernel32.dll');
+final kernel32Lib = HybridCoreWindowsLibrary(_kernel32Lib);
 
 int LOBYTE(int w) => w & 0xff;
 int HIBYTE(int w) => (w >> 8) & 0xff;