+++ hybrid_core/hybrid_core_windows/lib/src/hybrid_core_windows_plugin.dart
@@ -1,9 +1,0 @@
-import 'package:hybrid_core_platform_interface/hybrid_core_platform_interface.dart';
-
-import 'os.dart';
-
-abstract class HybridCoreWindowsPlugin {
-  static void registerWith() {
-    OSPlatform.instance = WindowsPlatform();
-  }
-}