+++ hybrid_vision/hybrid_vision_android/example/lib/views/views.dart
@@ -1,0 +1,1 @@
+export 'barcode_detection_view.dart';