+++ hybrid_vision/hybrid_vision_darwin/lib/src/hybrid_vision_darwin_plugin.dart
@@ -1,9 +1,0 @@
-import 'package:hybrid_vision_platform_interface/hybrid_vision_platform_interface.dart';
-
-import 'barcode_detection.dart';
-
-abstract class HybridVisionDarwinPlugin {
-  static void registerWith() {
-    BarcodeDetectionPlatform.instance = DarwinBarcodeDetectionPlatform();
-  }
-}