+++ hybrid_core/hybrid_core_windows/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.5.2
+
+* Update `pubspec.lock`.
+* Move `WindowsCoreLibrary` to `HybridCoreWindowsLibrary`.
+
+## 0.5.1
+
+* Move `WindowsCorePlugin` to `HybridCoreWindowsPlugin`.
+
+## 0.5.0
+
+* Add example.
+* Move `HybridCoreWindowsPlugin` to `WindowsCorePlugin`.
+* Move `HybridCoreWindowsLibrary` to `WindowsCoreLibrary`.
+* Move `WindowsPlatform` to `WindowsImpl`.
+
 ## 0.3.4
 
 * Optimize project structure.